両方のモデルを使うようにする
